Fixed a bug on the Registration page
On the sign Up page, where you have the "Already have an account? Sign In", the black text color was same all through, hence  the 'color: #000;' was changed to 'color: #0ABAB5 bold;', so users can easily identify and understand it is clickable. also the color for the hover was added which is #c4c4c4;. (edited the css)